backend/pyproject.toml
[project]
name = "question-answer-backend"
version = "0.1.0"
description = "FastAPI backend for question answering with Langfuse tracking"
requires-python = ">=3.13"
dependencies = [
    "langfuse<3.0.0",
    "openai>=2.8.1",
    "anthropic>=0.40.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "azure-ai-inference>=1.0.0b9",
    "ipykernel>=7.1.0",
    "python-multipart>=0.0.9",
    "pymupdf>=1.24.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "pytest-cov>=4.1.0",
]

